+++ helpers_test.go
@@ -1,75 +1,62 @@
 package xlog
 
 import (
+	"html/template"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/require"
-	"github.com/emad-elsaid/xlog/markdown/text"
 )
 
-func TestBanner(t *testing.T) {
-	tcs := []struct {
-		name     string
-		path     string
-		content  string
-		expected string
-	}{
-		{
-			name:     "page in root and image is relative implicitly",
-			path:     "home",
-			content:  "![](image.jpg)",
-			expected: "/image.jpg",
-		},
-		{
-			name:     "page in root and image is relative explicitly",
-			path:     "home",
-			content:  "![](./image.jpg)",
-			expected: "/image.jpg",
-		},
-		{
-			name:     "page in root and image is relative explicitly in subdir",
-			path:     "home",
-			content:  "![](./images/image.jpg)",
-			expected: "/images/image.jpg",
-		},
-		{
-			name:     "page in subdir and image is relative implicitly",
-			path:     "posts/home",
-			content:  "![](image.jpg)",
-			expected: "/posts/image.jpg",
-		},
-		{
-			name:     "page in subdir and image is relative explicitly",
-			path:     "posts/home",
-			content:  "![](./image.jpg)",
-			expected: "/posts/image.jpg",
-		},
-		{
-			name:     "page in subdir and image is relative explicitly in subdir",
-			path:     "posts/home",
-			content:  "![](./images/image.jpg)",
-			expected: "/posts/images/image.jpg",
-		},
-		{
-			name:     "page in subdir and image is relative explicitly in parent",
-			path:     "posts/home",
-			content:  "![](../images/image.jpg)",
-			expected: "/images/image.jpg",
-		},
-	}
+// TestAgoBehavior tests that ago function behavior is preserved
+func TestAgoBehavior(t *testing.T) {
+	app := newTestApp()
 
-	for _, tc := range tcs {
-		t.Run(tc.name, func(t *testing.T) {
-			reader := text.NewReader([]byte(tc.content))
-			p := page{
-				name:       tc.path,
-				lastUpdate: time.Time{},
-				ast:        MarkdownConverter().Parser().Parse(reader),
-				content:    (*Markdown)(&tc.content),
-			}
+	app.config.Readonly = true
+	now := time.Now()
+	result := app.ago(now)
+	require.Equal(t, now.Format("Monday 2 January 2006"), result)
 
-			require.Equal(t, tc.expected, Banner(&p))
-		})
-	}
+	app.config.Readonly = false
+
+	recent := time.Now().Add(-500 * time.Millisecond)
+	result = app.ago(recent)
+	require.Contains(t, result, "Less than a second")
+
+	oneMinuteAgo := time.Now().Add(-1 * time.Minute)
+	result = app.ago(oneMinuteAgo)
+	require.Contains(t, result, "1 minute")
+
+	oneHourAgo := time.Now().Add(-1 * time.Hour)
+	result = app.ago(oneHourAgo)
+	require.Contains(t, result, "1 hour")
+
+	oneDayAgo := time.Now().Add(-24 * time.Hour)
+	result = app.ago(oneDayAgo)
+	require.Contains(t, result, "1 day")
+}
+
+// TestIsFontAwesome tests that IsFontAwesome function behavior is preserved
+func TestIsFontAwesome(t *testing.T) {
+	require.True(t, IsFontAwesome("fa-solid"), "Expected 'fa-solid' to be FontAwesome")
+	require.True(t, IsFontAwesome("fa-regular"), "Expected 'fa-regular' to be FontAwesome")
+	require.True(t, IsFontAwesome("fa-brands"), "Expected 'fa-brands' to be FontAwesome")
+	require.False(t, IsFontAwesome("not-fa"), "Expected 'not-fa' to not be FontAwesome")
+	require.False(t, IsFontAwesome(""), "Expected empty string to not be FontAwesome")
+}
+
+// TestDir tests that dir function behavior is preserved
+func TestDir(t *testing.T) {
+	require.Equal(t, "", dir(""), "Expected empty string for empty path")
+	require.Equal(t, "", dir("."), "Expected empty string for '.'")
+	require.Equal(t, "", dir("file.txt"), "Expected empty string for 'file.txt'")
+	require.Equal(t, "dir", dir("dir/file.txt"), "Expected 'dir' for 'dir/file.txt'")
+	require.Equal(t, "a/b", dir("a/b/c.txt"), "Expected 'a/b' for 'a/b/c.txt'")
+}
+
+// TestRaw tests that raw function behavior is preserved
+func TestRaw(t *testing.T) {
+	input := "<div>test</div>"
+	result := raw(input)
+	require.Equal(t, template.HTML(input), result, "Expected HTML to match input")
 }
